# apps/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: White House Village Manager API
  version: 0.1.0
  description: |
    Contrato da API de gestão da White House Village.

    Este arquivo é a **fonte da verdade**: é mergeado antes da implementação, e é
    contra ele que o painel e os testes são escritos. Pertence ao agente
    `tech-lead`; os demais propõem alterações, não editam.

    ## Campo desconhecido no corpo é erro, nunca silêncio

    Toda requisição com corpo **recusa** propriedade que o schema não declara,
    com `422 VALIDATION_ERROR` e o nome do campo em `details`. Não é purismo: um
    corpo com campo errado que responde `200` é a pior resposta possível — o
    cliente acha que gravou, o servidor não gravou nada, e não há erro em lugar
    nenhum para alguém investigar. Foi exatamente assim que
    `PATCH /reservations/{id}` com `{"guests": 6}` respondeu `200` e manteve
    `guests_count` em 4 (medido, 26/08/2026).

    A regra vale para `POST`, `PUT` e `PATCH`. Em `PATCH`, campo **ausente** não
    muda nada e campo com `null` limpa — ausência é silêncio combinado; nome
    desconhecido não é.

    ## Um dado, um nome

    Uma informação tem **um** nome em todo o contrato: o mesmo na entrada, na
    saída e na coluna do banco. Número de hóspedes é `guests_count` em pedido de
    orçamento, criação, edição, remarcação e resposta — não `guests` em metade
    delas.
servers:
  - url: http://localhost:8080/api/v1
    description: Desenvolvimento
...
paths:
...
  /auth/login:
    post:
      tags: [Autenticação]
      summary: Autentica e devolve o par de tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
      responses:
        "200":
          description: Autenticado
          headers:
            Set-Cookie:
              description: |
                `wh_refresh=<token>; HttpOnly; SameSite=Lax; Path=/api/v1/auth`
                (+ `Secure` em produção), válido por 30 dias. O refresh nunca vai
                no corpo por padrão — fora do cookie `httpOnly` ele fica ao
                alcance de qualquer XSS.
              schema: { type: string }
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      access_token: { type: string }
                      expires_in: { type: integer, example: 900 }
                      user: { $ref: "#/components/schemas/Usuario" }
        "401":
          description: |
            `INVALID_CREDENTIALS` — a resposta é **idêntica** para e-mail
            inexistente, senha errada e e-mail bloqueado por tentativas
            (5 erros em 15 min bloqueiam por 15 min). Diferenciar aqui seria
            entregar de graça a lista de quem tem conta.
          $ref: "#/components/responses/Erro"
...
  /users/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Usuários]
      summary: Detalhe do usuário
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Usuario" }
        "404": { $ref: "#/components/responses/Erro" }
    put:
      tags: [Usuários]
      summary: Substitui o usuário
      description: |
        Substituição integral: campo opcional ausente volta ao padrão
        (`phone` fica nulo, `active` volta a `true`). `password` ausente **mantém
        a senha atual** — PUT não é motivo para derrubar a senha de ninguém.
        Para alteração parcial use `PATCH`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/UsuarioAtualizar"
                - type: object
                  required: [name, email, role_id]
      responses:
        "200":
          description: Atualizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Usuario" }
        "403":
          description: |
            `FORBIDDEN` — `role_id` de perfil mais poderoso que o do ator, ou
            ator sem `roles:editar`. Atribuir papel é administrar acesso.
          $ref: "#/components/responses/Erro"
        "404": { $ref: "#/components/responses/Erro" }
        "409":
          description: "`EMAIL_IN_USE` — o e-mail já pertence a outro usuário."
          $ref: "#/components/responses/Erro"
        "422": { $ref: "#/components/responses/Erro" }
    patch:
      tags: [Usuários]
      summary: Atualiza parcialmente o usuário
      description: |
        Só o que veio no corpo muda. Campo **ausente** não é tocado; campo com
        `null` limpa (vale para `phone`) — a distinção que o `Opt[T]` carrega.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UsuarioAtualizar" }
      responses:
        "200":
          description: Atualizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Usuario" }
        "403":
          description: |
            `FORBIDDEN` — `role_id` de perfil mais poderoso que o do ator, ou
            ator sem `roles:editar`. Atribuir papel é administrar acesso.
          $ref: "#/components/responses/Erro"
        "404": { $ref: "#/components/responses/Erro" }
        "409":
          description: "`EMAIL_IN_USE` — o e-mail já pertence a outro usuário."
          $ref: "#/components/responses/Erro"
        "422": { $ref: "#/components/responses/Erro" }
    delete:
      tags: [Usuários]
      summary: Desativa o usuário (soft delete)
      description: |
        Marca `active = false` e revoga todas as famílias de sessão. Não apaga a
        linha: reserva, oportunidade e auditoria apontam para o usuário, e
        histórico com autor órfão não é histórico. Reativar é `PATCH` com
        `active: true`.
      responses:
        "204": { description: Desativado }
        "404": { $ref: "#/components/responses/Erro" }
        "422":
          description: |
            `VALIDATION_ERROR` — o usuário autenticado não pode desativar a si
            mesmo, nem o último usuário ativo de perfil `admin` (ficaria uma
            instalação sem ninguém capaz de configurar).
          $ref: "#/components/responses/Erro"
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  responses:
    Erro:
      description: Erro com código estável
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  code:
                    type: string
                    description: |
                      Vocabulário fechado e estável — o front reage ao `code`,
                      nunca ao texto de `message`. Espelhado em
                      `internal/platform/apperr`: acrescentar código aqui obriga
                      a acrescentar lá, e o contrário também.
                    enum: [VALIDATION_ERROR, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, DATE_CONFLICT,
                           MIN_STAY_NOT_MET, CAPACITY_EXCEEDED, DISCOUNT_ABOVE_LIMIT, HOLD_EXPIRED,
                           IDEMPOTENCY_MISMATCH, RATE_LIMITED, INVALID_CREDENTIALS, TOKEN_INVALID,
                           TOKEN_REUSED, EMAIL_IN_USE, ROLE_IMMUTABLE, ROLE_IN_USE,
                           RATE_NOT_FOUND, POLICY_IMMUTABLE, CODE_IN_USE, RESOURCE_IN_USE,
                           INVALID_STATE_TRANSITION, RESERVATION_NOT_CANCELLABLE,
                           UNIT_NOT_AVAILABLE, COMPOSITION_INCOMPLETE, HOLD_LIMIT_REACHED,
                           INTERNAL]
                  message: { type: string }
                  details: { type: object, additionalProperties: true }
  schemas:
...
    Usuario:
      type: object
      properties:
        id:      { type: string, format: uuid }
        name:    { type: string }
        email:   { type: string, format: email }
        phone:   { type: [string, "null"], example: "+5585999990000" }
        role_id: { type: string, format: uuid }
        role:
          type: string
          example: corretor
          description: |
            `code` do perfil. **Não é enum fechado** — perfil é dado, criável em
            `/roles`; front que faça `switch` por este valor quebra no primeiro
            perfil novo. Para decidir o que mostrar, use `permissions`.
        role_name: { type: string, example: Corretor }
        broker_id:
          type: [string, "null"]
          format: uuid
          description: Só leitura. Preenchido quando a conta é a de um corretor.
        active:  { type: boolean }
        last_login_at: { type: [string, "null"], format: date-time }
        created_at:    { type: string, format: date-time }
...
    UsuarioAtualizar:
      type: object
      description: |
        Corpo do `PATCH` (tudo opcional) e base do `PUT`. Campo **ausente** não
        muda nada; campo com `null` limpa — é a distinção que o `Opt[T]` do Go
        carrega e que um `map[string]any` perderia.
      properties:
        name:     { type: string, minLength: 2 }
        email:    { type: string, format: email }
        password: { type: string, format: password, minLength: 8, description: "Trocar a senha revoga todas as sessões do usuário." }
        role_id:
          type: string
          format: uuid
          description: |
            Muda o perfil; passa a valer na requisição seguinte, sem novo login.

            Atribuir papel é ato de administração de acesso: exige
            `roles:editar` (não basta `usuarios:editar`), e o papel alvo só é
            aceito se a matriz dele for **subconjunto** da do ator — ninguém
            concede o que não tem, senão editar usuário vira escalada de
            privilégio em um passo. Fora disso: `403 FORBIDDEN`.

            Trocar o **próprio** `role_id` é recusado com `422
            VALIDATION_ERROR`, pela mesma razão que ninguém se desativa: o
            caminho para uma instalação sem administrador não pode passar por
            um clique distraído.
        phone:    { type: [string, "null"] }
        active:   { type: boolean, description: "`false` bloqueia o login e revoga as sessões — mesmo efeito do `DELETE`." }


    # ─────────────────────────── Inventário ───────────────────────────
...
security:
  - bearerAuth: []